+++ proxy/body_test.go
@@ -8,7 +8,7 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	http_proxy "github.com/AndreaCostanzo1/http-proxy/pkg"
+	http_proxy "github.com/AndreaCostanzo1/http-proxy/proxy"
 )
 
 func TestSetBody(t *testing.T) {
