+++ proxy/interceptors_test.go
@@ -5,7 +5,7 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	http_proxy "github.com/AndreaCostanzo1/http-proxy/pkg"
+	http_proxy "github.com/AndreaCostanzo1/http-proxy/proxy"
 )
 
 func TestWithGenericInterceptor(t *testing.T) {
